Update README.md
Given a pointer to the head of a linked list and a specific position, determine the data value at that position. Count backwards from the tail node. The tail is at postion 0, its parent is at 1 and so on.

Example
head refers to 3 ->2->1->null

Each of the data values matches its distance from the tail. The value  is at the desired position.